# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NSE Finance API",
    "description": "Python FastAPI replacement for NestJS 9 backend. Keeps MySQL finance_app (localhost:3306) and frontend compatibility GET /stocks?duration=... and GET /stock-exchange/Nse. New domain: portfolios / positions / trades / price_snapshots.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/stocks": {
      "get": {
        "tags": [
          "stocks"
        ],
        "summary": "List Stocks",
        "operationId": "list_stocks_stocks_get",
        "parameters": [
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "tweek|lweek|tmonth|lmonth|tyear|lyear",
              "title": "Duration"
            },
            "description": "tweek|lweek|tmonth|lmonth|tyear|lyear"
          },
          {
            "name": "load",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "if 'true', enrich each row with LTP/52W",
              "title": "Load"
            },
            "description": "if 'true', enrich each row with LTP/52W"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Stocks Stocks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "stocks"
        ],
        "summary": "Create stock (legacy compat)",
        "operationId": "create_stock_stocks_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Stock Stocks Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/stocks/{stock_id}": {
      "patch": {
        "tags": [
          "stocks"
        ],
        "summary": "Update stock",
        "operationId": "update_stock_stocks__stock_id__patch",
        "parameters": [
          {
            "name": "stock_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StockUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Stock Stocks  Stock Id  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "stocks"
        ],
        "summary": "Delete stock",
        "operationId": "delete_stock_stocks__stock_id__delete",
        "parameters": [
          {
            "name": "stock_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Stock Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Stock Stocks  Stock Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StockCreate": {
        "properties": {
          "username": {
            "type": "string",
            "minLength": 1,
            "title": "Username"
          },
          "stockName": {
            "type": "string",
            "minLength": 1,
            "title": "Stockname"
          },
          "price": {
            "type": "string",
            "minLength": 1,
            "title": "Price"
          },
          "quantity": {
            "type": "string",
            "minLength": 1,
            "title": "Quantity"
          },
          "type": {
            "$ref": "#/components/schemas/TransactionType",
            "description": "buy | sell"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "username",
          "stockName",
          "price",
          "quantity",
          "type"
        ],
        "title": "StockCreate"
      },
      "StockUpdate": {
        "properties": {
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "stockName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stockname"
          }
        },
        "type": "object",
        "title": "StockUpdate"
      },
...
      "TransactionType": {
        "type": "string",
        "enum": [
          "buy",
          "sell"
        ],
        "title": "TransactionType"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}